requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
numpy
scikit-learn
sentence_transformers
langchain_huggingface
chromadb
faiss-cpu
langchain_chroma
langchain_groq
langchain_Core
fastapi
uvicorn
langserve
sse_starlette
# UI
streamlit==1.31.0
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
# # Agents and tools
langchainhub
duckduckgo-search